x402-data-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
web3==7.2.0
python-dotenv==1.0.1
